ENT-289 Improve accounting for inactive courses
We were missing a corner case where courses stop being returned for a catalog,
in which case we will need to mark them as inactive.

This change intoduces a lightweight audit summary of the course updates which
will be used in subsequent pushes to fully account for course state changes.