Refactor session termination code.
This commit refactors the work done in a previous commit which added
logic to the enterprise_login_required decorator which forced the user
to have a fresh session each time they make a request to the decorated view.
This refactor moves the fresh session logic out into a separate decorator so
that it can be applied independently of the enterprise_login_required decorator.
We only need to apply the fresh session logic to views which will be used as
launch points for users coming over to our site from a third-party site with
SSO.

ENT-454