ENT-581 Record one completion event per enrollment record
Previously, we were allowing multiple completion events per enrollment to be
sent in case something had changed that was relevant to send again. We were using
the timestamp to determine whether an even has changed. However, for audit/self paced
enrollments, the completed timestamp is incorrect and reflects the time the job runs.
This is because the grades api doesn't return the date the grade was passing. So for
now, we will send just one event per enrollment since there is no strong need to do
otherwise.