Add Degreed as a new integrated channel.
In addition to adding Degreed, we refactor the existing
code infrastructure for integrated channels to match the
following structure:

Integration Channel
    - Exporters
        - Learner Data
        - Course Metadata
    - Transmitters
        - Learner Data
        - Course Metadata
    - API Client
    - Models
        - Global Integration Channel Configuration
        - Integration Channel Enterprise Customer Configuration
        - Catalog Data Transmission Audit
        - Learner Data Transmission Audit

- Exporters

An exporter is essentially a serializer (I would have gone with this name but we
were already using 'Exporter' so I decided to go with it).

An exporter will yield serialized data given the interface method export.
We choose to use yield as the output of this interface because it allows
us to loop over this serialized data at a rate that does not surpass the
limit that some integration channel may have for receiving data on the other end.

Exporters may do whatever they must to gather data, including possibly calling
other APIs (course API, certificate API, etc.), but make a promise through
the export method to yield some serialized data, to be used by a transmitter.

- Transmitters

A transmitter transmits exported (read: serialized) data to the integration channel
using the API client. It uses the transmit interface method which takes a single
payload, which is an Exporter object (so that it can be exported (serialized)).
It then delegates the actual transmission of the payload to the API client.

- API Client

The API Client for an integration channel is the specialized client that knows
how the integration channel expects to be tickled, typically through a RESTful interface.

The API client is responsible for, usually, making the RESTful call given a serialized
payload that it receives from some transmitter. It will handle any authentication
details that need to be dealt with for the given integration channel's interface.

- Models

Each integration channel may or may not require certain models. If it does not need
to build its own, it may delegate any model requirements to the base integration channel models.

- Global Integrated Channel Configuration

A global integration channel configuration defines some globally required
configuration for a given integration channel, usually, for example, defining
API paths, provider IDs, and so on. This configuration is then used throughout
the integration channel's code, common to all Enterprise Customers that use
this integration channel.

- Integration Channel Enterprise Customer Configuration

An integration channel enterprise customer configuration is a enterprise
customer-specific configuration model for the given integration channel;
it defines some details that may differ among enterprise customers for
the given integration channel.

- Catalog Data Transmission Audit

A catalog data transmission audit is a model used to keep track of course
metadata that is transmitted to a given integration channel. It may differ
according to the integration channel in question, so different integration
channels may define their own.

- Learner Data Transmission Audit

A learner data transmission audit is a model used to keep track of learner
data that is transmitted to a given integration channel. It may differ
according to the integration channel in question, so different integration
channels may define their own.