ENT-1006 Fixing requirements snafoo before I do anything else
I wasn't able to run make upgrade and make requirements in order to
run manage.py to make migrations, so I went down a rabbit hole that led
to pinning readme_renderer in doc.in and html5lib in the release.in files.
This is all to keep us at bleach==1.4, which is what edx-platform is pinned to.