Remove a GET side-effect from consent flow
1.2.4 attempted to fix the side effect in HandleConsentEnrollment by
removing the view and calling the code as a method inside the post
call. It was ignoring the redirect url it was being passed and
ran all requests through this new method instead. This broke coupon
redemption and redirects from courseware.

This fixes by only using the new method when the redirect URL is a
specific string, thus allowing all previous pathways to work as they did
before.

ENT-1307