Change api to allow client to send in boolean field (for now still in request params, later as patch body)
This change will allow the client side code to be optimized to use boolean for marked_done rather than a string.
This is incremental and in near future we will update it to use patch body and process this field as boolean. right now it gets a string 'true' or 'false' due to string based payload

Fixes: ENT-3047