fix: handle case where LearnerCreditEnterpriseCourseEnrollment already exists
When calling enroll_learners_in_courses (a.k.a. the bulk enrollment
endpoint), gracefully handle the case when a subsidized enrollment (for
learner credit 2) is being re-created after being revoked previously.

Before this change, the existing LearnerCreditEnterpriseCourseEnrollment
object would not be found (due to transaction_id changing), then the
view would fatally fail to create a new one due to a unique constraint
on the `enterprise_course_enrollment_id` field.  After this change, it
finds the existing object and updates a couple fields on it to re-use
it.

Bump version to 3.65.3.

ENT-7204