feat: add enterprise_catalog_query to output of catalog list endpoint
This is needed to support robust provisioning, editing, and displaying
of Learner Credit Plans linked to "predefined" catalogs. The previous
behavior was to rely on reading the catalog title and use string
manipulation (fragile) to infer the catalog query type. Exposing the
query ID allows the LC provisioning tool to more directly understand the
query used.

ENT-7922