fix: Do NOT skip validating intentions which have NO related catalog queries
When testing in prod, I found that Verizon Skills Forward was skipped
because it only had a custom/curated catalog (catalog query was NULL).
Greatest() in MySQL returns NULL if any input is NULL (unlike
PostgreSQL), so adding Coalesce() with an arbitrarily old date is
necessary to mimic the PostgreSQL behavior.

ENT-9941