Refactor  add foreign key to ecu part 3 (#2341)
Part 3 of refactoring described in #2338 : adding a foreign key to the `EnterpriseCustomerUser` (ECU) table.
This adds a management command that backfills `user_fk` values in the ECU table from `user_id`, in batches.
Additionally, it adds logic to the model's `save()` method that copies `user_id` to `user_fk` when a record is created or updated.