Refactor  add foreign key to ecu part 5 (#2345)
Part of https://github.com/openedx/edx-enterprise/pull/2338, see there for extended description.
The overall goal is to add a foreign key to the EnterpriseCustomerUser that points to User.
Since this is a large table, a series of PRs and migrations was designed to do this without incurring performance problems.
This migration is the last part, adding the actual foreign key to the previously added and indexed `user_fk` field.
Before applying this to a very large table, to avoid downtime, you **MUST** make sure that no `user_fk` entries are NULL (but the column must remain nullable).
In order to make that happen, deploy https://github.com/openedx/edx-enterprise/pull/2341 first and run the management task defined there.