fix: Fix a bug in the social auth redirect logic.
Previously the test was calling `called_once` on a mock, which is a
non-existent function so the test would pass even when it wasn't
supposed to.  In python 3.12 it fails which led us to fining the bug in
our code.  The check `handle_redirect_after_social_auth_login` was not
properly handling the redirect when a user has more than one enterprise
association and is trying to enroll in a course.

This commit fixes the tests to use the correct assert and then updates
the code to fix the bug so that tests pass as expected.