build: Switch back to upstream tincan.
The tincan library now supports Python 3.x. We originally forked this
repo because it was not able make that update in a timely manner for us
to hit our upgrade deadlines.  But now that it has, there is no reason
for us to keep our own fork of the repo.

The changes made in the fork have mostly been to get the repo to be easy
to maintain by the open edx community and not to the baseline logic or
capabilities of the repo.

A diff can be found here: https://github.com/RusticiSoftware/TinCanPython/compare/3.x...openedx:TinCanPython:master